refactor(frontend): the appliance editors and the upload pipeline out of the panel
The documents, metadata and parts editors move to `panel-asset-editors.ts`; the
size pre-check, progress bar, failure messages and file picker behind them move
to `panel-upload.ts`. The editors call into the pipeline and never the other way
round, so `setAssetError` and `errorAlert` sit with the pipeline that raises
them rather than above it.

Five duplications went with the move, each written once now:

- `fileCard` — an appliance document and a part's file were two copies of the
  same card, already drifting in how they built their subtitle.
- `filePicker` + `UPLOAD_ACCEPT` — both upload buttons spelled the accepted MIME
  list out by hand, so editing one left the other behind.
- `documentSchema` — the name + URL grid, written identically in the add form
  and in an existing link's inline editor.
- `mutateDocuments` — add / update / remove were three copies of one envelope
  (unsaved appliance mutates the working copy, saved one calls the service and
  refreshes from the response, either way an error lands on the form).
- `entryBox` — the framed head with its confirm-first Remove, shared by a
  metadata entry and a part.

`PanelHost` grows by four: `_attachNotePreview` (the parts editor's note
previews must be registered for disposal, so the panel stays their only
constructor) and the three upload fields the pipeline owns.

Verified byte-identical: the built bundle's template literals are the same set
(one fewer occurrence — the two `entryBox` heads are now one), and the rendered
shadow tree for an appliance drawer carrying documents, metadata and parts,
including every live `ha-form`'s schema and data, diffs clean against the same
render before the move.
